hf-vision-demo/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.44.0
torch==2.2.0
torchvision==0.17.0
opencv-python==4.8.0.76
python-multipart==0.0.6
numpy==1.26.0
pillow==10.0.0
aiofiles==23.1.0
accelerate==0.24.1
timm==1.0.20
